refactor[SP2-11]: Tags module
